requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
google-genai>=1.0.0
chromadb==0.5.15
sentence-transformers==3.1.1
rank-bm25==0.2.2
python-dotenv==1.0.1
httpx==0.27.2
PyYAML==6.0.2
python-multipart==0.0.9
pytest==8.3.3
pytest-asyncio==0.24.0
streamlit==1.38.0
requests==2.32.3
